the basic functionality for 8.1
restructured ReceivingStates to act as a real StateMachine (transition on input)
distinction between general inputstream and dataInputStream in H2Stream

When the receiver has the correct class the receive method is called which in turn calls the subclassspecific process method. after that receive is used to save some state information and handle the completion of the communication if necissary.